config: add UrlResolver, separate data server config from resolver
