+++ htsget-config/src/regex_resolver/aws.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 /// Configuration for the htsget server.
 #[derive(Deserialize, Serialize, Debug, Clone, Default)]
 #[serde(default)]
-pub struct AwsS3DataServer {
+pub struct S3Resolver {
   pub bucket: String,
   pub cors_allow_credentials: bool,
   pub cors_allow_origin: String,
